Made it so that integer is always int64. had to change indKX etc to be floats so that they can be multiplied by rhos and phis.

+++ run_dir/run/parameter.py
@@ -6,10 +6,9 @@ import cupy as np
 ###########################
 #Tf = np.float32
 #Tc = np.complex64
-#Ti = np.int16
 Tf = np.float64
 Tc = np.complex128
-Ti = np.int32
+Ti = np.int64
 
 ################
 ### Ensemble ###
